requirements.txt
langgraph
langchain-openai
langchain-core
langchain-text-splitters
psycopg2-binary
python-dotenv
python-docx
ipython
